% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{continuous.schmidt.decompose}
\alias{continuous.schmidt.decompose}
\title{Perform a continuous Schmidt decomposition}
\usage{
continuous.schmidt.decompose(bifunc, x1lo, x1hi, x2lo, x2hi, nbx1 = 100,
  nbx2 = 100, keep = min(10, nbx1, nbx2))
}
\description{
Perform a continuous Schmidt decomposition
}
