% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{continuous_function_interpolate}
\alias{continuous_function_interpolate}
\title{Interpolate values of functions.}
\usage{
continuous_function_interpolate(xarr, yarr, x)
}
\arguments{
\item{xarr}{a vector of x (sorted)}

\item{yarr}{a vector of y}

\item{x}{given value of x}
}
\value{
interpolated value of y
}
\description{
Interpolate values of functions.
}
