% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{discretize.continuous.bipartitefunc}
\alias{discretize.continuous.bipartitefunc}
\title{Making a discretized tensor for a continuous function}
\usage{
discretize.continuous.bipartitefunc(bifunc, x1lo, x1hi, x2lo, x2hi,
  nbx1 = 100, nbx2 = 100)
}
\arguments{
\item{bifunc}{bipartitite continuous wavefunction}

\item{x1lo}{lower limit of \code{x1}}

\item{x1hi}{upper limit of \code{x1}}

\item{x2lo}{lower limit of \code{x2}}

\item{x2hi}{upper limit of \code{x2}}

\item{nbx1}{number of discretized x1 (default: 100)}

\item{nbx2}{number of discretized x2 (default: 100)}
}
\value{
discretized tensor for Schmidt decomposition
}
\description{
Making a discretized tensor for a continuous function
}
