% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{entanglement.entropy}
\alias{entanglement.entropy}
\title{Calculate the entanglement entropy given the calculate Schmidt modes.}
\usage{
entanglement.entropy(modes)
}
\arguments{
\item{calculated}{Schmidt modes}
}
\value{
entanglement entropy
}
\description{
Calculate the entanglement entropy given the calculate Schmidt modes.
}
