% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{interpolated.continuous.function}
\alias{interpolated.continuous.function}
\title{Lambda function of the interpolated continous function.}
\usage{
interpolated.continuous.function(xarr, yarr)
}
\arguments{
\item{xarr}{a vector of x (sorted)}

\item{yarr}{a vector of y}
}
\value{
interpolated lambda function
}
\description{
Lambda function of the interpolated continous function.
}
