% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{negativity}
\alias{negativity}
\title{Calculate the negativity given the calculate Schmidt modes.}
\usage{
negativity(modes)
}
\arguments{
\item{modes}{Schmidt modes}
}
\value{
negativity
}
\description{
Calculate the negativity given the calculate Schmidt modes.
}
