% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{participation.ratio}
\alias{participation.ratio}
\title{Calculate the participation ratio given the calculate Schmidt modes.}
\usage{
participation.ratio(modes)
}
\arguments{
\item{calculated}{Schmidt modes}
}
\value{
participation ratio
}
\description{
Calculate the participation ratio given the calculate Schmidt modes.
}
