% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{reduced.denmat}
\alias{reduced.denmat}
\title{Get reduced density matrix}
\usage{
reduced.denmat(bipartite.qubits, keep.dim = 1)
}
\arguments{
\item{keep.dim}{dimension to keep (default: 1)}

\item{bipartitite.qubits}{tensor of bipartite systems}
}
\value{
reduced density matrix
}
\description{
Get reduced density matrix
}
