% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{schmidt.decompose}
\alias{schmidt.decompose}
\title{Perform Schmidt decomposition}
\usage{
schmidt.decompose(bipartite.qubits)
}
\arguments{
\item{red.dm}{reduced density matrix}
}
\value{
Schmidt modes, including the eigenvalues, and eigenvectors of both subsystems of the modes
}
\description{
Perform Schmidt decomposition
}

