% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{entanglement.entropy}
\alias{entanglement.entropy}
\title{Calculate the entanglement entropy given the calculate Schmidt modes.}
\usage{
entanglement.entropy(modes)
}
\arguments{
\item{modes}{Schmidt modes}
}
\value{
entanglement entropy
}
\description{
Calculate the entanglement entropy given the calculate Schmidt modes.
}
\examples{
singlet<- matrix(c(0, sqrt(0.7), sqrt(0.3), 0), byrow = TRUE, nrow = 2)
modes<- schmidt.decompose(singlet)
entanglement.entropy(modes)

}
