% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{negativity}
\alias{negativity}
\title{Calculate the negativity given the calculate Schmidt modes.}
\usage{
negativity(modes)
}
\arguments{
\item{modes}{Schmidt modes}
}
\value{
negativity
}
\description{
Calculate the negativity given the calculate Schmidt modes.
}
\examples{
singlet<- matrix(c(0, sqrt(0.7), sqrt(0.3), 0), byrow = TRUE, nrow = 2)
modes<- schmidt.decompose(singlet)
negativity(modes)

}
