% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{reduced.denmat}
\alias{reduced.denmat}
\title{Get reduced density matrix}
\usage{
reduced.denmat(bipartite.qubits, keep.dim = 1)
}
\arguments{
\item{bipartite.qubits}{tensor of bipartite systems}

\item{keep.dim}{dimension to keep (default: 1)}
}
\value{
reduced density matrix
}
\description{
Get reduced density matrix
}
\examples{
singlet<- matrix(c(0, sqrt(0.7), sqrt(0.3), 0), byrow = TRUE, nrow = 2)
reduced.denmat(singlet)

}
